Extract sidebar/layout logic from AppState into an extension (#71)
